Remove old STR handler
